+++ src/__tests__/client.init.test.ts
@@ -1,6 +1,6 @@
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
-import RyskSDK from 'src'
+import CitrexSDK from 'src'
 import { Environment } from 'src/enums'
 import { privateKey } from 'vitest/utils'
 
@@ -17,19 +17,19 @@ vi.mock('viem', async () => {
   }
 })
 
-describe('The RyskSDK', () => {
+describe('The CitrexSDK', () => {
   beforeEach(() => {
     fetchMock.resetMocks()
   })
 
   it('should initialise correctly with only required params', () => {
-    const Client = new RyskSDK(privateKey)
+    const Client = new CitrexSDK(privateKey)
 
     expect(Client).toMatchSnapshot()
   })
 
   it('should initialise correctly with all config parameters passed', () => {
-    const Client = new RyskSDK(privateKey, {
+    const Client = new CitrexSDK(privateKey, {
       debug: true,
       environment: Environment.TESTNET,
       rpc: 'https://test-rpc.quiknode.pro',
@@ -42,7 +42,7 @@ describe('The RyskSDK', () => {
   it('should POST the referral code if running on mainnet', async () => {
     fetchMock.mockResponse(JSON.stringify({ success: true }))
 
-    new RyskSDK(privateKey, { environment: Environment.MAINNET })
+    new CitrexSDK(privateKey, { environment: Environment.MAINNET })
 
     await vi.waitFor(() => {
       if (!fetchMock.mock.calls.length) throw new Error('Awaiting call...')
@@ -50,7 +50,7 @@ describe('The RyskSDK', () => {
 
     expect(fetchMock.mock.lastCall).toMatchInlineSnapshot(`
       [
-        "https://api.rysk.finance/v1/vault/referral",
+        "https://api.citrex.markets/v1/vault/referral",
         {
           "body": "{"account":"0xb47B0b1e44B932Ae9Bb01817E7010A553A965Ea8","code":"kickflip"}",
           "method": "POST",
@@ -62,7 +62,7 @@ describe('The RyskSDK', () => {
   it('should handle referral failures gracefully', async () => {
     fetchMock.mockReject(new Error('An unknown error occurred'))
 
-    const Client = new RyskSDK(privateKey, { debug: true, environment: Environment.MAINNET })
+    const Client = new CitrexSDK(privateKey, { debug: true, environment: Environment.MAINNET })
 
     await vi.waitFor(() => {
       if (Client.logs.length !== 3) throw new Error('Awaiting call...')