+++ src/__tests__/client.orders.test.ts
@@ -1,12 +1,12 @@
 import { recoverTypedDataAddress } from 'viem'
 import { beforeEach, describe, expect, it, vi } from 'vitest'
 
-import RyskSDK from 'src'
+import CitrexSDK from 'src'
 import EIP712 from 'src/ABI/EIP712'
 import { TimeInForce, OrderType } from 'src/enums'
 import { privateKey, address, marketFOKOrder, customOrder, replacementOrder } from 'vitest/utils'
 
-describe('The RyskSDK', () => {
+describe('The CitrexSDK', () => {
   beforeEach(() => {
     vi.useRealTimers()
     vi.restoreAllMocks()
@@ -18,7 +18,7 @@ describe('The RyskSDK', () => {
     it('should allow a user to successfully place a market order', async () => {
       fetchMock.mockResponse(JSON.stringify(marketFOKOrder))
 
-      const Client = new RyskSDK(privateKey)
+      const Client = new CitrexSDK(privateKey)
 
       await vi.advanceTimersByTimeAsync(10000)
 
@@ -53,7 +53,7 @@ describe('The RyskSDK', () => {
       ).toEqual(address)
       expect(call).toMatchInlineSnapshot(`
         [
-          "https://api.staging.rysk.finance/v1/order",
+          "https://api.staging.citrex.markets/v1/order",
           {
             "body": "{"expiration":1712421770000,"nonce":1709829780000000,"price":"3536200000000000000000","quantity":"1000000000000000","signature":"0xab9a2fc1f10e2d0e81f7a3831f35c2dce96008281774ad2a85c0dc02ecd47935024c20f9b9241495b46d9007c5520b12f15dd186fdaf846777a8dc287ac148f31b","account":"0xb47B0b1e44B932Ae9Bb01817E7010A553A965Ea8","isBuy":true,"orderType":2,"productId":1002,"subAccountId":1,"timeInForce":1}",
             "method": "POST",
@@ -68,7 +68,7 @@ describe('The RyskSDK', () => {
     it('should allow a user to successfully place an order with all args passed (limit order)', async () => {
       fetchMock.mockResponse(JSON.stringify(customOrder))
 
-      const Client = new RyskSDK(privateKey)
+      const Client = new CitrexSDK(privateKey)
 
       const result = await Client.placeOrder({
         expiration: 1800000000000,
@@ -104,7 +104,7 @@ describe('The RyskSDK', () => {
       ).toEqual(address)
       expect(call).toMatchInlineSnapshot(`
         [
-          "https://api.staging.rysk.finance/v1/order",
+          "https://api.staging.citrex.markets/v1/order",
           {
             "body": "{"expiration":1800000000000,"nonce":123,"price":"3450000000000000000000","quantity":"1000000000000000","signature":"0x2cd5bdcf7586bb615e3ea7e0b696a6cdc31a564dbab23114475705003081a9cd18e1f9634421ffb650d4992d71009482bb061dc326ff3a42cd013d8eec082cb31c","account":"0xb47B0b1e44B932Ae9Bb01817E7010A553A965Ea8","isBuy":true,"orderType":0,"productId":1002,"subAccountId":1,"timeInForce":2}",
             "method": "POST",
@@ -119,7 +119,7 @@ describe('The RyskSDK', () => {
     it('should allow a user to specify slippage', async () => {
       fetchMock.mockResponse(JSON.stringify(marketFOKOrder))
 
-      const Client = new RyskSDK(privateKey)
+      const Client = new CitrexSDK(privateKey)
 
       await Client.placeOrder({
         isBuy: true,
@@ -146,7 +146,7 @@ describe('The RyskSDK', () => {
     })
 
     it('should return an error for a market order with no priceIncrement', async () => {
-      const Client = new RyskSDK(privateKey)
+      const Client = new CitrexSDK(privateKey)
 
       const result = await Client.placeOrder({
         isBuy: true,
@@ -166,7 +166,7 @@ describe('The RyskSDK', () => {
     it('should handle an error during the order process', async () => {
       fetchMock.mockResponse(JSON.stringify({ error: 'A known error occurred' }))
 
-      const Client = new RyskSDK(privateKey)
+      const Client = new CitrexSDK(privateKey)
 
       const result = await Client.placeOrder({
         isBuy: true,
@@ -187,7 +187,7 @@ describe('The RyskSDK', () => {
     it('should handle an unknown error', async () => {
       fetchMock.mockReject(new Error('An unknown error occurred'))
 
-      const Client = new RyskSDK(privateKey)
+      const Client = new CitrexSDK(privateKey)
 
       const result = await Client.placeOrder({
         isBuy: true,
@@ -212,7 +212,7 @@ describe('The RyskSDK', () => {
         .mockOnce(JSON.stringify({ error: 'A known error occurred' }))
         .mockReject(new Error('An unknown error occurred'))
 
-      const Client = new RyskSDK(privateKey)
+      const Client = new CitrexSDK(privateKey)
 
       const result = await Client.placeOrders([
         {
@@ -285,7 +285,7 @@ describe('The RyskSDK', () => {
     it('should allow a user to successfully cancel an order', async () => {
       fetchMock.mockResponse(JSON.stringify({ success: true }))
 
-      const Client = new RyskSDK(privateKey)
+      const Client = new CitrexSDK(privateKey)
 
       const result = await Client.cancelOrder(
         '0x3505c6219b1f51cf216e432b153f8637c1fa9342520bd7c780bd80dafe0eed94',
@@ -309,7 +309,7 @@ describe('The RyskSDK', () => {
       ).toEqual(address)
       expect(call).toMatchInlineSnapshot(`
         [
-          "https://api.staging.rysk.finance/v1/order",
+          "https://api.staging.citrex.markets/v1/order",
           {
             "body": "{"account":"0xb47B0b1e44B932Ae9Bb01817E7010A553A965Ea8","orderId":"0x3505c6219b1f51cf216e432b153f8637c1fa9342520bd7c780bd80dafe0eed94","productId":1002,"subAccountId":1,"signature":"0x98f70a0f2849a47f01e28b3ae9c0cbfd1f27f7209ca1ce71a524caf6100df88513ec831a84d37fb0913065d74e7fd8cdd4094d9f10e5db7a9a823d2b91f643a81c"}",
             "method": "DELETE",
@@ -322,7 +322,7 @@ describe('The RyskSDK', () => {
     it('should handle an error during the cancel process', async () => {
       fetchMock.mockResponse(JSON.stringify({ error: 'A known error occurred', success: false }))
 
-      const Client = new RyskSDK(privateKey)
+      const Client = new CitrexSDK(privateKey)
 
       const result = await Client.cancelOrder(
         '0x3505c6219b1f51cf216e432b153f8637c1fa9342520bd7c780bd80dafe0eed94',
@@ -340,7 +340,7 @@ describe('The RyskSDK', () => {
     it('should handle an unknown error', async () => {
       fetchMock.mockReject(new Error('An unknown error occurred'))
 
-      const Client = new RyskSDK(privateKey)
+      const Client = new CitrexSDK(privateKey)
 
       const result = await Client.cancelOrder(
         '0x3505c6219b1f51cf216e432b153f8637c1fa9342520bd7c780bd80dafe0eed94',
@@ -361,7 +361,7 @@ describe('The RyskSDK', () => {
         .mockOnce(JSON.stringify({ error: 'A known error occurred', success: false }))
         .mockReject(new Error('An unknown error occurred'))
 
-      const Client = new RyskSDK(privateKey)
+      const Client = new CitrexSDK(privateKey)
 
       const result = await Client.cancelOrders([
         ['0x3505c6219b1f51cf216e432b153f8637c1fa9342520bd7c780bd80dafe0eed94', 1002],
@@ -391,7 +391,7 @@ describe('The RyskSDK', () => {
     it('should allow a user to successfully cancel all orders for a product', async () => {
       fetchMock.mockResponse(JSON.stringify({ success: true }))
 
-      const Client = new RyskSDK(privateKey)
+      const Client = new CitrexSDK(privateKey)
 
       const result = await Client.cancelOpenOrdersForProduct(1002)
       const call = fetchMock.mock.calls[0]
@@ -411,7 +411,7 @@ describe('The RyskSDK', () => {
       ).toEqual(address)
       expect(call).toMatchInlineSnapshot(`
         [
-          "https://api.staging.rysk.finance/v1/openOrders",
+          "https://api.staging.citrex.markets/v1/openOrders",
           {
             "body": "{"account":"0xb47B0b1e44B932Ae9Bb01817E7010A553A965Ea8","productId":1002,"subAccountId":1,"signature":"0xdf4d9439710977064a05cc2fe92df9c516615e3b49ca359165ada4f8eb19e8b738e2cafc2ea8c04cee08afeffcfc7bfeb962becdb376e2db8c87717bf0c8b2011b"}",
             "method": "DELETE",
@@ -424,7 +424,7 @@ describe('The RyskSDK', () => {
     it('should handle an error during the cancel process', async () => {
       fetchMock.mockResponse(JSON.stringify({ error: 'A known error occurred', success: false }))
 
-      const Client = new RyskSDK(privateKey)
+      const Client = new CitrexSDK(privateKey)
 
       const result = await Client.cancelOpenOrdersForProduct(1002)
 
@@ -439,7 +439,7 @@ describe('The RyskSDK', () => {
     it('should handle an unknown error', async () => {
       fetchMock.mockReject(new Error('An unknown error occurred'))
 
-      const Client = new RyskSDK(privateKey)
+      const Client = new CitrexSDK(privateKey)
 
       const result = await Client.cancelOpenOrdersForProduct(1002)
 
@@ -454,7 +454,7 @@ describe('The RyskSDK', () => {
     it('should allow a user to cancel and replace an order', async () => {
       fetchMock.mockResponse(JSON.stringify(replacementOrder))
 
-      const Client = new RyskSDK(privateKey)
+      const Client = new CitrexSDK(privateKey)
 
       const result = await Client.cancelAndReplaceOrder(replacementOrder.id, {
         isBuy: true,
@@ -486,7 +486,7 @@ describe('The RyskSDK', () => {
       ).toEqual(address)
       expect(call).toMatchInlineSnapshot(`
         [
-          "https://api.staging.rysk.finance/v1/order/cancel-and-replace",
+          "https://api.staging.citrex.markets/v1/order/cancel-and-replace",
           {
             "body": "{"idToCancel":"0x08d4079c501e5fbb2153c7fe785ea4648ffcdac411d93511edcb5b18aecc158f","newOrder":{"expiration":1712421760000,"nonce":1709829760000000,"price":"3455000000000000000000","quantity":"1000000000000000","signature":"0xd8601fb78e8a22a37b53c8c5365b45d53344832524276faeaa890bcd3e38e50746a1d08b38008931c81c8b186ef02da49e172540ffcee7b8554534d4687ec6831c","account":"0xb47B0b1e44B932Ae9Bb01817E7010A553A965Ea8","isBuy":true,"orderType":1,"productId":1002,"subAccountId":1,"timeInForce":0}}",
             "method": "POST",
@@ -522,7 +522,7 @@ describe('The RyskSDK', () => {
         JSON.stringify({ ...replacementOrder, expiry: 1800000000000, nonce: 123 }),
       )
 
-      const Client = new RyskSDK(privateKey)
+      const Client = new CitrexSDK(privateKey)
 
       const result = await Client.cancelAndReplaceOrder(replacementOrder.id, {
         expiration: 1800000000000,
@@ -556,7 +556,7 @@ describe('The RyskSDK', () => {
       ).toEqual(address)
       expect(call).toMatchInlineSnapshot(`
         [
-          "https://api.staging.rysk.finance/v1/order/cancel-and-replace",
+          "https://api.staging.citrex.markets/v1/order/cancel-and-replace",
           {
             "body": "{"idToCancel":"0x08d4079c501e5fbb2153c7fe785ea4648ffcdac411d93511edcb5b18aecc158f","newOrder":{"expiration":1800000000000,"nonce":123,"price":"3455000000000000000000","quantity":"1000000000000000","signature":"0xe85e19f4ac1f9389e9d4e89be1bdac251c207feb8c816d815a7c0c5caa1404782abc79e01a56ec949b536d2a50ef3fb9cc844e5069a32d80721cfc975a4e10cd1c","account":"0xb47B0b1e44B932Ae9Bb01817E7010A553A965Ea8","isBuy":true,"orderType":1,"productId":1002,"subAccountId":1,"timeInForce":0}}",
             "method": "POST",
@@ -590,7 +590,7 @@ describe('The RyskSDK', () => {
     it('should handle an error during the cancel process', async () => {
       fetchMock.mockResponse(JSON.stringify({ error: 'A known error occurred', success: false }))
 
-      const Client = new RyskSDK(privateKey)
+      const Client = new CitrexSDK(privateKey)
 
       const result = await Client.cancelAndReplaceOrder(replacementOrder.id, {
         isBuy: true,
@@ -610,7 +610,7 @@ describe('The RyskSDK', () => {
     it('should handle an unknown error', async () => {
       fetchMock.mockReject(new Error('An unknown error occurred'))
 
-      const Client = new RyskSDK(privateKey)
+      const Client = new CitrexSDK(privateKey)
 
       const result = await Client.cancelAndReplaceOrder(replacementOrder.id, {
         isBuy: true,