+++ src/constants/__mocks__/api.ts
@@ -1,8 +1,8 @@
 import type { Environment } from 'src/types'
 
 const API_URL = {
-  mainnet: 'https://api.rysk.finance/v1',
-  testnet: 'https://api.staging.rysk.finance/v1',
+  mainnet: 'https://api.citrex.markets/v1',
+  testnet: 'https://api.staging.citrex.markets/v1',
 } as Record<Environment, string>
 
 export default API_URL