+++ src/constants/chains.ts
@@ -1,11 +1,11 @@
 import type { Environment } from 'src/types'
 import type { Chain } from 'viem'
 
-import { arbitrum, arbitrumSepolia } from 'viem/chains'
+import { sei, seiTestnet } from 'viem/chains'
 
 const CHAINS: Record<Environment, Chain> = {
-  mainnet: arbitrum,
-  testnet: arbitrumSepolia,
+  mainnet: sei,
+  testnet: seiTestnet,
 }
 
 export default CHAINS